Added test for successfully shutting down validator and starting up leader
